Remove this unnecessary cast to "int".

+++ src/mines/Board.java
@@ -84,7 +84,7 @@ public class Board extends JPanel {
         i = 0;
         while (i < mines) {
 
-            position = (int) (allCells * random.nextInt());
+            position =  (allCells * random.nextInt());
 
             if ((position < allCells) &&
                 (field[position] != COVERED_MINE_CELL)) {
